requirements.txt
# الأساسيات
azure-functions>=1.15.0,<2.0.0
pillow>=10.1.0,<11.0.0

# التخزين
azure-storage-blob>=12.17.0,<13.0.0

# المراقبة
opencensus-ext-azure>=1.1.11,<2.0.0
# أو:
# azure-monitor-opentelemetry>=1.0.0

# التبعيات الاختيارية
python-dotenv>=1.0.0  # لإدارة المتغيرات البيئية